+++ prisma/schema.prisma
@@ -23,7 +23,7 @@ model Person {
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
 
-  Response Response[]
+  Response Result[]
 }
 
 model Church {
@@ -35,7 +35,7 @@ model Church {
   updatedAt DateTime @updatedAt
 }
 
-model Response {
+model Result {
   id       String   @id @default(cuid())
   personId  String  @unique
   person    Person   @relation(fields: [personId], references: [id])
